+++ src/lib/services/searchService.ts
@@ -1,17 +1,18 @@
 import { useQuery } from "@tanstack/react-query";
-import { useEffect, useState } from "react";
+import { useEffect, useMemo, useState } from "react";
 
 import {
   useBaseApiRoute,
   useCelatoneApp,
   useGetAddressType,
 } from "lib/app-provider";
-import type { ContractAddr } from "lib/types";
+import type { Addr, ContractAddr, Option } from "lib/types";
 import { isBlock, isCodeId, isTxHash } from "lib/utils";
 
 import { useBlockDetailsQuery } from "./blockService";
 import { useCodeDataByCodeId } from "./codeService";
 import { queryContract } from "./contract";
+import { useAddressByICNSName } from "./nameService";
 import { useTxData } from "./txService";
 
 export type SearchResultType =
@@ -22,11 +23,21 @@ export type SearchResultType =
   | "Proposal ID"
   | "Block";
 
-// TODO: Add Proposal ID, ICNS query
+export interface ResultMetadata {
+  icns: {
+    address: Option<Addr>;
+  };
+}
+
+// TODO: Add Proposal ID
 export const useSearchHandler = (
   keyword: string,
   resetHandlerStates: () => void
-): { results: SearchResultType[]; isLoading: boolean } => {
+): {
+  results: SearchResultType[];
+  isLoading: boolean;
+  metadata: ResultMetadata;
+} => {
   const [debouncedKeyword, setDebouncedKeyword] = useState(keyword);
   const {
     chainConfig: {
@@ -54,12 +65,28 @@ export const useSearchHandler = (
   );
   const { data: blockData, isLoading: blockLoading } =
     useBlockDetailsQuery(debouncedKeyword);
+  const { data: icnsAddressData, isLoading: icnsAddressLoading } =
+    useAddressByICNSName(debouncedKeyword);
+
   const txDataLoading = isTxHash(debouncedKeyword) && txLoading;
   const codeDataLoading = isWasm && isCodeId(debouncedKeyword) && codeLoading;
   const contractDataLoading = isWasm && contractLoading;
   const blockDataLoading = isBlock(debouncedKeyword) && blockLoading;
   const isAddr =
     addressType === "user_address" || addressType === "contract_address";
+
+  const addressResult = useMemo(() => {
+    if (isAddr) {
+      return contractData ? "Contract Address" : "Wallet Address";
+    }
+
+    return (
+      icnsAddressData?.address &&
+      (icnsAddressData.addressType === "contract_address"
+        ? "Contract Address"
+        : "Wallet Address")
+    );
+  }, [isAddr, contractData, icnsAddressData]);
 
   useEffect(() => {
     const timeoutId = setTimeout(() => {
@@ -71,7 +98,7 @@ export const useSearchHandler = (
 
   return {
     results: [
-      isAddr && (contractData ? "Contract Address" : "Wallet Address"),
+      addressResult,
       txData && "Transaction Hash",
       codeData && "Code ID",
       blockData && "Block",
@@ -80,6 +107,12 @@ export const useSearchHandler = (
       txDataLoading ||
       codeDataLoading ||
       contractDataLoading ||
-      blockDataLoading,
+      blockDataLoading ||
+      icnsAddressLoading,
+    metadata: {
+      icns: {
+        address: icnsAddressData?.address,
+      },
+    },
   };
 };